Modify api routing
