+++ config/routes.rb
@@ -8,9 +8,11 @@ Rails.application.routes.draw do
   # Defines the root path route ("/")
   # root "posts#index"
   Rails.application.routes.draw do
-    root "news#index"
   
-    resources :news, only: [:index, :show, :create, :update, :destroy]
+    namespace :api do
+      namespace :v1 do
+        resources :news, only: [:index, :show, :create, :update, :destroy]
+      end
+    end
   end
-  
 end
